feat(gopwt): support unary exprs in assert. (+, -, ^, &, <-)

+++ main.go
@@ -62,7 +62,9 @@ func doMain() error {
 	}
 
 	tmpGopath, importpath, err := translator.Translate(flag.Arg(0))
-	defer os.RemoveAll(tmpGopath)
+	if os.Getenv("GOPWT_DEBUG") == "" { // NOTE: don't remove translated files on GOPWT_DEBUG=1
+		defer os.RemoveAll(tmpGopath)
+	}
 	if err != nil {
 		return err
 	}
